requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# miccai20mil/research_mil/evaluators/utils.py: 13
MulticoreTSNE == 0.1

# miccai20mil/research_mil/evaluators/utils.py: 185
# miccai20mil/research_mil/loaders/datasets.py: 13
Pillow == 8.0.1

# miccai20mil/research_mil/test.py: 5
# miccai20mil/research_mil/train.py: 7
PyYAML == 5.3.1

# miccai20mil/research_mil/evaluators/utils.py: 10,12,15
# miccai20mil/research_mil/loaders/datasets.py: 8
matplotlib == 3.2.1

# miccai20mil/research_mil/evaluators/base_eval.py: 8
# miccai20mil/research_mil/evaluators/global_tester.py: 8
# miccai20mil/research_mil/evaluators/globaltaskmil_test.py: 3
# miccai20mil/research_mil/evaluators/utils.py: 3
# miccai20mil/research_mil/loaders/datasets.py: 5
# miccai20mil/research_mil/train.py: 8
# miccai20mil/research_mil/trainers/base_trainer.py: 8
# miccai20mil/research_mil/trainers/global_trainer.py: 8
# miccai20mil/research_mil/trainers/globaltaskmil_trainer.py: 6
# miccai20mil/research_mil/utils/mil.py: 1
# miccai20mil/research_mil/utils/misc.py: 3
numpy == 1.19.2

# miccai20mil/research_mil/loaders/datasets.py: 6
openslide_python == 1.1.1

# miccai20mil/research_mil/loaders/datasets.py: 7
pandas == 1.0.3

# miccai20mil/research_mil/evaluators/utils.py: 14
scikit_learn == 0.23.0

# miccai20mil/research_mil/evaluators/utils.py: 16
seaborn == 0.10.1

# miccai20mil/research_mil/train.py: 14
tensorboardX == 2.0

# miccai20mil/research_mil/evaluators/base_eval.py: 2,3
# miccai20mil/research_mil/evaluators/global_tester.py: 2,3
# miccai20mil/research_mil/evaluators/globaltaskmil_test.py: 2
# miccai20mil/research_mil/evaluators/utils.py: 2
# miccai20mil/research_mil/loaders/__init__.py: 1,2
# miccai20mil/research_mil/loaders/datasets.py: 11
# miccai20mil/research_mil/models/__init__.py: 2,3
# miccai20mil/research_mil/models/backbones/resnetv2.py: 1,2,3
# miccai20mil/research_mil/models/pooling/methods.py: 1,2,3
# miccai20mil/research_mil/train.py: 8
# miccai20mil/research_mil/trainers/base_trainer.py: 2,3
# miccai20mil/research_mil/trainers/global_trainer.py: 2,3
# miccai20mil/research_mil/trainers/globaltaskmil_trainer.py: 2,3,4,5,10
# miccai20mil/research_mil/utils/misc.py: 2
torch == 1.6.0

# miccai20mil/research_mil/evaluators/globaltaskmil_test.py: 6
# miccai20mil/research_mil/evaluators/utils.py: 9
# miccai20mil/research_mil/loaders/datasets.py: 10,254
# miccai20mil/research_mil/trainers/globaltaskmil_trainer.py: 9
torchvision == 0.7.0

# miccai20mil/research_mil/evaluators/globaltaskmil_test.py: 5
# miccai20mil/research_mil/evaluators/utils.py: 8
# miccai20mil/research_mil/loaders/datasets.py: 4
# miccai20mil/research_mil/trainers/globaltaskmil_trainer.py: 8
tqdm == 4.46.0
